board with color & fixed positions color-circles
